third_party/WebKit/LayoutTests/svg/custom/svg-parse-overflow-3-expected.txt

CONSOLE ERROR: line 14: Error: <rect> attribute height: Expected length, "1050000000000000…".
